EcommerceApi/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-filter==21.1
django-js-asset==2.0.0
djangorestframework==3.13.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.1.0
importlib-metadata==4.11.2
Markdown==3.3.6
Pillow==9.0.1
PyJWT==1.7.1
pytz==2021.3
sqlparse==0.4.2
whitenoise==6.0.0
zipp==3.7.0
